Make YamlDb an interface and provide an implementation
